refactor: separate atom instance data
